requirements.txt
pydantic >= 2.10.2
rich >= 13.9.4
curl_cffi >= 0.7.3
pyyaml >= 6.0.2
